requirements.txt
torch==2.6.0+cu118
torchvision==0.21.0+cu118
torchaudio==2.6.0+cu118
triton==3.2.0

transformers==4.57.1
timm==1.0.21
tokenizers==0.22.1
safetensors==0.6.2
huggingface-hub==0.36.0

numpy==2.1.2
scipy==1.15.3
sympy==1.13.1
einops==0.8.1
torchinfo==1.8.0

opencv-python==4.12.0.88
pillow==11.3.0
matplotlib==3.10.7

tqdm==4.67.1
requests==2.32.5
regex==2025.10.23
PyYAML==6.0.3
